Add support for arbitrary number of biasing cycles per voltage.
Normally for each voltage the selected tests are applied in order. In order
to gather more statistics allow for more than one set of tests per voltage.
Currently anything more than 1 will probably break the fitting with the
parametric model. All cycles of each voltages must be averaged in order for
the model to properly be fitted.