Make repetitions repeat pairs of voltages instead of singular ones.
Biasing repetitions should repeat pairs of voltages instead of single ones.
As the code was structured during each cycle only a single polarity voltage
was repeated. For example for the pairs [(1,-1), (1.1, -1.1), (1.2, -1.2)]
the correct sequence should be 1,-1,1,-1,...(N times),... instead of
1,1,1,1,...(N times), -1,-1,-1,-1,...(N times). To do so the voltage
sequence table is constructed in beforehand so that the voltage are
repeated in pairs correctly when the routine runs.