Do an initialisation read step before running the first cycle.
The initialisation step allows us to save the number of cycles and pulses
in the initial tag in the form of MPF_INIT_CX_PY_s where X the number of
cycles and Y the number of pulses. This makes processing easier when
multiple cycles must be averaged before fitted to the model.