Update the resistance plot to account for multiple cycles
Since we are now supporting multiple cycles per voltage pair the plotting
algorithm needed a thorough overhaul to account for this change. The
resistance array is now a numpy.ndarray instead of python list holding the
data of each cycle as a separate row. Each row holds the entirety of the
resistance data for a single cycle. At the end the columns are averaged and
plotted as before.

The initialisation step (see commits 822254e and 4ecb9e9) is required to
provide for all the necessary counters, namely the number of cycles, the
number of points per cycle and the total number of voltage pairs used
during testing.  This allows us to allocate the relevant buffers that hold
the resistance, voltage and pulse width data as ndarrays are fixed size.
Once the buffers are allocated a line index is used to track the number of
lines processed. Another counter, the cycle index, tracks the row of the
resistance buffer in which the value should be written. The value is
written to buffer[cycle][line]. Once all the values of a single cycle and
voltage are written the line index is updated to point to the next buffer
index. A separate counter counts the voltage pairs that have been
processed. If the current cycle is the last the cycle counter is reset and
the voltage pair index is increased. In any case (whether this is the last
cycle or not) the line index is set to
2 x (voltage idx) x (pulses per cycle).

Once everything has been processed resistances are averaged across all
cycles and plotted.