Properly decouple the CT blocks from a single ParameterFit run.
Instead of processing the FF blocks in pairs as it was before it is done
for every single FF block (single voltage). This complicates the
procession of counters but allows the CT blocks to be parsed correctly and
be assigned to the correct resistance value. To do this the previous values
of cycleIdx and voltIdx is also required, since the modification of these
values always happens in the FF blocks (which are guaranteed to be
present) but the followup blocks (of the same) cycle must now the (now
altered) values. So the previous cycleIdx and voltIdx must be preserved.
This will probably be the case for the RET blocks as well.