Add a noise widget and extract retention data from the raw arrays
In order to incorporate a noise widget we need both the bias (ie.
PROGRAM) and retention (ie. READ) data. The retention values are now
extracted into a suitable array. The need to be further split in order to
accomodate for measurements during (or after) positive and negative bias
before any actual noise processing can be done.